[MOV] mail: move '_tracking_value_format' to tracking mixin
Currently, tracking value model has a '_tracking_value_format' helper
for formatting that calls '_tracking_value_format_model' that is
model-specific. This one exists notably to fetch model fields_get
for fields information, and allow model customization (see notably
account.return model).

In this commit we split this computation, to move the generic
formatting helper on mail.message, as it is called when computing
store information for chatter. The model specific one is moved
in the track mixin, to join other methods defined on models that
perform value tracking.

Task-5935695 ([mail, various] Cleanup and test tracking usage)
Prepares Task-3645865 ([mail] In-body tracking)
