+++ addons/mail/models/mail_track_mixin.py
@@ -4,7 +4,7 @@ import typing
 
 from datetime import datetime
 
-from odoo import fields, models
+from odoo import api, fields, models
 from odoo.exceptions import MissingError
 from odoo.tools import clean_context, ormcache
 
@@ -381,3 +381,111 @@ class MailTrackMixin(models.AbstractModel):
             value, False, col_name, col_info,
         )
         return {**tracking_values, 'field_info': field_info}
+
+    # track value formatting
+    # ------------------------------------------------------
+
+    @api.model
+    def _tracking_value_format_model(self, tracking_values: BaseModel) -> list[ValuesType]:
+        """ Return structured formatted data to be used by chatter to display
+        tracking values on a single model. Order it based on ascending sequence
+        then field name. Property fields are always last.
+
+        :returns: for each tracking value in self, their formatted display
+          values given as a dict;
+        :rtype: list[ValuesType]
+        """
+        if not tracking_values:
+            return []
+
+        # fetch model-based information
+        if self._name not in ('mail.thread', 'mail.track.mixin'):
+            tracked_fields = self.fields_get(tracking_values.field_id.mapped('name'), attributes={'digits', 'string', 'type'})
+            model_sequence_info = dict(self._mail_track_order_fields(tracked_fields))
+        else:
+            tracked_fields, model_sequence_info = {}, {}
+
+        # generate sequence of trackings
+        fields_sequence_map = dict(
+            {
+                tracking.field_info['name']: tracking.field_info.get('sequence', 100)
+                for tracking in tracking_values.filtered('field_info')
+            },
+            **model_sequence_info,
+        )
+        # generate dict of field information, if available
+        fields_col_info = (
+            tracking.field_id.ttype != 'properties'
+            and tracked_fields.get(tracking.field_id.name)
+            or {
+                'string': tracking.field_info['desc'] if tracking.field_info else tracking_values.env._('Unknown'),
+                'type': tracking.field_info['type'] if tracking.field_info else 'char',
+            } for tracking in tracking_values
+        )
+
+        def sort_tracking_info(tracking_info_tuple):
+            tracking = tracking_info_tuple[0]
+            field_name = tracking.field_id.name or (tracking.field_info['name'] if tracking.field_info else 'unknown')
+            return (
+                fields_sequence_map.get(field_name, 100),
+                tracking.field_id.ttype == 'properties',
+                field_name,
+            )
+
+        formatted = [
+            {
+                'id': tracking.id,
+                'fieldInfo': {
+                    'changedField': col_info['string'],
+                    'currencyId': tracking.currency_id.id,
+                    'floatPrecision': col_info.get('digits'),
+                    'fieldType': col_info['type'],
+                    'isPropertyField': tracking.field_id.ttype == 'properties',
+                },
+                'newValue': self._format_display_value(tracking, col_info['type'], new=True)[0],
+                'oldValue': self._format_display_value(tracking, col_info['type'], new=False)[0],
+            }
+            for tracking, col_info in sorted(zip(tracking_values, fields_col_info), key=sort_tracking_info)
+        ]
+        return formatted
+
+    @api.model
+    def _format_display_value(self, trackings, field_type: str, new: bool = True) -> str:
+        """ Format value of 'mail.tracking.value', according to the field type.
+
+        :param str field_type: Odoo field type;
+        :param bool new: if True, display the 'new' value. Otherwise display
+          the 'old' one.
+        """
+        field_mapping = {
+            'boolean': ('old_value_integer', 'new_value_integer'),
+            'date': ('old_value_datetime', 'new_value_datetime'),
+            'datetime': ('old_value_datetime', 'new_value_datetime'),
+            'char': ('old_value_char', 'new_value_char'),
+            'float': ('old_value_float', 'new_value_float'),
+            'integer': ('old_value_integer', 'new_value_integer'),
+            'monetary': ('old_value_float', 'new_value_float'),
+            'text': ('old_value_text', 'new_value_text'),
+        }
+
+        result = []
+        for record in trackings:
+            value_fname = field_mapping.get(
+                field_type, ('old_value_char', 'new_value_char')
+            )[bool(new)]
+            value = record[value_fname]
+
+            if field_type in {'integer', 'float', 'char', 'text', 'monetary'}:
+                result.append(value)
+            elif field_type in {'date', 'datetime'}:
+                if not record[value_fname]:
+                    result.append(value)
+                elif field_type == 'date':
+                    result.append(fields.Date.to_string(value))
+                else:
+                    result.append(f'{value}Z')
+            elif field_type == 'boolean':
+                result.append(bool(value))
+            else:
+                result.append(value)
+        return result
