+++ addons/mail/models/mail_tracking_value.py
@@ -1,12 +1,5 @@
 # -*- coding: utf-8 -*-
-from __future__ import annotations
-
-import typing
-
 from odoo import fields, models
-
-if typing.TYPE_CHECKING:
-    from odoo.api import ValuesType
 
 
 class MailTrackingValue(models.Model):
@@ -66,125 +59,3 @@ class MailTrackingValue(models.Model):
             return model_field and not model_field.groups
 
         return self.filtered(has_free_access)
-
-    def _tracking_value_format(self) -> list[ValuesType]:
-        """ Return structured formatted data to be used by chatter to display
-        tracking values. It is organized by model.
-
-        :return: for each tracking value in self, their formatted display
-          values given as a dict;
-        :rtype: list[ValuesType]
-        """
-        model_map = {}
-        for tracking in self:
-            model = tracking.field_id.model or tracking.mail_message_id.model
-            model_map.setdefault(model, self.browse())
-            model_map[model] += tracking
-        formatted = []
-        for model, trackings in model_map.items():
-            formatted += trackings._tracking_value_format_model(model)
-        return formatted
-
-    def _tracking_value_format_model(self, model: str) -> list[ValuesType]:
-        """ Return structured formatted data to be used by chatter to display
-        tracking values on a single model. Order it based on ascending sequence
-        then field name. Property fields are always last.
-
-        :returns: for each tracking value in self, their formatted display
-          values given as a dict;
-        :rtype: list[ValuesType]
-        """
-        if not self:
-            return []
-
-        # fetch model-based information
-        if model:
-            TrackedModel = self.env[model]
-            tracked_fields = TrackedModel.fields_get(self.field_id.mapped('name'), attributes={'digits', 'string', 'type'})
-            model_sequence_info = dict(TrackedModel._mail_track_order_fields(tracked_fields)) if model else {}
-        else:
-            tracked_fields, model_sequence_info = {}, {}
-
-        # generate sequence of trackings
-        fields_sequence_map = dict(
-            {
-                tracking.field_info['name']: tracking.field_info.get('sequence', 100)
-                for tracking in self.filtered('field_info')
-            },
-            **model_sequence_info,
-        )
-        # generate dict of field information, if available
-        fields_col_info = (
-            tracking.field_id.ttype != 'properties'
-            and tracked_fields.get(tracking.field_id.name)
-            or {
-                'string': tracking.field_info['desc'] if tracking.field_info else self.env._('Unknown'),
-                'type': tracking.field_info['type'] if tracking.field_info else 'char',
-            } for tracking in self
-        )
-
-        def sort_tracking_info(tracking_info_tuple):
-            tracking = tracking_info_tuple[0]
-            field_name = tracking.field_id.name or (tracking.field_info['name'] if tracking.field_info else 'unknown')
-            return (
-                fields_sequence_map.get(field_name, 100),
-                tracking.field_id.ttype == 'properties',
-                field_name,
-            )
-
-        formatted = [
-            {
-                'id': tracking.id,
-                'fieldInfo': {
-                    'changedField': col_info['string'],
-                    'currencyId': tracking.currency_id.id,
-                    'floatPrecision': col_info.get('digits'),
-                    'fieldType': col_info['type'],
-                    'isPropertyField': tracking.field_id.ttype == 'properties',
-                },
-                'newValue': tracking._format_display_value(col_info['type'], new=True)[0],
-                'oldValue': tracking._format_display_value(col_info['type'], new=False)[0],
-            }
-            for tracking, col_info in sorted(zip(self, fields_col_info), key=sort_tracking_info)
-        ]
-        return formatted
-
-    def _format_display_value(self, field_type: str, new: bool = True) -> str:
-        """ Format value of 'mail.tracking.value', according to the field type.
-
-        :param str field_type: Odoo field type;
-        :param bool new: if True, display the 'new' value. Otherwise display
-          the 'old' one.
-        """
-        field_mapping = {
-            'boolean': ('old_value_integer', 'new_value_integer'),
-            'date': ('old_value_datetime', 'new_value_datetime'),
-            'datetime': ('old_value_datetime', 'new_value_datetime'),
-            'char': ('old_value_char', 'new_value_char'),
-            'float': ('old_value_float', 'new_value_float'),
-            'integer': ('old_value_integer', 'new_value_integer'),
-            'monetary': ('old_value_float', 'new_value_float'),
-            'text': ('old_value_text', 'new_value_text'),
-        }
-
-        result = []
-        for record in self:
-            value_fname = field_mapping.get(
-                field_type, ('old_value_char', 'new_value_char')
-            )[bool(new)]
-            value = record[value_fname]
-
-            if field_type in {'integer', 'float', 'char', 'text', 'monetary'}:
-                result.append(value)
-            elif field_type in {'date', 'datetime'}:
-                if not record[value_fname]:
-                    result.append(value)
-                elif field_type == 'date':
-                    result.append(fields.Date.to_string(value))
-                else:
-                    result.append(f'{value}Z')
-            elif field_type == 'boolean':
-                result.append(bool(value))
-            else:
-                result.append(value)
-        return result
